# Conflicts:
#	src/sdl12/Srb2SDL-vc10.vcxproj
#	src/sdl12/Srb2SDL-vc9.vcproj
#	src/sdl12/macosx/Srb2mac.pbproj/project.pbxproj
#	src/sdl12/macosx/Srb2mac.xcodeproj/project.pbxproj

+++ src/sdl/i_system.c
@@ -2044,8 +2044,8 @@ static void I_ShutdownTimer(void)
 //
 tic_t I_GetTime (void)
 {
-	static Uint32 basetime = 0;
-		   Uint32 ticks = SDL_GetTicks();
+	static Uint64 basetime = 0;
+		   Uint64 ticks = SDL_GetTicks();
 
 	if (!basetime)
 		basetime = ticks;
